Split game objects from board state
